backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
python-multipart
pandas
apscheduler
PyJWT
numpy==1.23.5
pandas-ta==0.3.14b0